+++ deconversation/__init__.py
@@ -1,4 +1,4 @@
-__version__ = "0.0.2"
+__version__ = "0.0.3"
 #from .core import deconverse
 def __getattr__(name):
     if name == "deconverse":